More work on build for Smart Tool
Ported TMC22xx driver here and added option to user a SERCOM to provide the UART
Added separate AnalogIn module for SAMC21 because it is so different from the SAME51
Fixed up differences in DMA controller programming
Note: SmartTool configuration still doesn't build, but it is close. Expansion1 builds tested and working.