Fix heater faults in the presence of extrusion temperature boost
The old logic was calculating temperature rise across an arbitrary
sequence of temperature boost and de-boosts, potentially from different
starting temperatures, which doesn't make sense physically. This
resulted in temperatureRisingTooSlowly faults at random. Some of these
faults would even report a negative temperature rise (with fully
functioning hardware and good tunings) due to calculating the rise from
different starting temperatures.

The new logic will still trigger faults for complete hardware failures,
because the mode will stay in HeaterMode::heating as the sensor reading
continues to fall which means the values for calculating the heating
rate will not be reset.

Also add a missing InterruptCriticalSectionLocker, which is already
present in the RepRapFirmware version of this file.